Include a coherent nework generator by jklais
Adds a modified version of user jklails' coherent network generator. It is modified in order to work in Python 3 (and networkx 2) and additional modifications will be made as necessary. It is licensed under the MIT license, so I think this is fine, but I may need to include a copywrite notice, I'm unsure currently.