Edit sparse training algorithm
Removing and replacing of edges is now done globally, according to the algorithm described at https://arxiv.org/pdf/1902.05967.pdf
This allows edges to be removed from one layer and then added in another, adding more flexibility. Note that some elements of the previous algorithm have been kept, such as the initialisation making the layers with fewer nodes more dense, and removing the bottom few percent of edges (rather than the suggestion in the above paper for estimating the number of edges without performing a sort, since sorting is nowhere near being a bottleneck currently).