Fixed bottleneck in back-propagation
Fixed the performance bottleneck during back-propagation, this was actually coming from taking the transpose of a CSR matrix, turning it into a CSC matrix, and then multiplying on the left by it, which is very inefficient. Now the transpose is calculated once and stored separately, increasing memory usage during training but increasing performance as much as 50%. Fixes #5.