Added Selective Weight Decay
Add the Selective Weight Decay method, and also the Sparse Networks from Scratch method (in place of "global"), and the RigL method (in place of "Local"). For the second two, the methods are fairly similar to the methods already being implemented, and produce fairly similar results. Selective Weight Decay is a true pruning technique, and not a sparse training technique, so produces better results at the cost of higher computational complexity. The hope is that the results are more robust against starting conditions and so produce more reliable results. The papers references are
https://arxiv.org/pdf/1907.04840 - Sparse Networks From Scratch
https://arxiv.org/abs/1911.11134 - RigL
https://arxiv.org/abs/2011.10520 - Selective Weight Decay