Added new incoherence model
Incoherence is now introduced by allowing the user to explicitly specify which layers each layer can connect to, giving much more control over the shape of the network whilst still allowing incoherence to appear.
Other changes are removing AdaDelta as a method, since it hadn't been updated to allow incoherence and didn't show better performance than momentum, so was deemed not worth keeping. Also now there is annealing of the learning rate, which seems fairly hit and miss whether it seems useful or not so will likely be tweaked in the future.
Random number seeding has been updated, the outcome is the same however the recommended method is to create a generator rather than repeatedly seeding the global one.